+++ crates/ruma-federation-api/src/transactions/edu.rs
@@ -207,7 +207,8 @@ pub struct DeviceListUpdateContent {
     pub device_id: DeviceIdBox,
 
     /// The public human-readable name of this device. Will be absent if the device has no name.
-    pub device_display_name: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub device_display_name: Option<String>,
 
     /// An ID sent by the server for this update, unique for a given user_id.
     pub stream_id: UInt,
@@ -227,13 +228,13 @@ pub struct DeviceListUpdateContent {
 }
 
 impl DeviceListUpdateContent {
-    /// Creates a new `DeviceListUpdateContent` with default `prev_id`, `deleted`
-    /// and `keys` fields.
-    pub fn new(user_id: UserId, device_id: DeviceIdBox, name: String, stream_id: UInt) -> Self {
+    /// Creates a new `DeviceListUpdateContent` with default `prev_id`, `deleted`,
+    /// `device_display_name` and `keys` fields.
+    pub fn new(user_id: UserId, device_id: DeviceIdBox, stream_id: UInt) -> Self {
         Self {
             user_id,
             device_id,
-            device_display_name: name,
+            device_display_name: None,
             stream_id,
             prev_id: vec![],
             deleted: None,
@@ -328,7 +329,7 @@ mod test {
                 keys
             }) if user_id == &user_id!("@john:example.com")
                 && device_id == &device_id!("QBUAZIFURK")
-                && device_display_name == "Mobile"
+                && device_display_name.as_deref() == Some("Mobile")
                 && stream_id == &uint!(6)
                 && prev_id.is_empty()
                 && deleted == &Some(false)
@@ -338,6 +339,43 @@ mod test {
         assert_eq!(
             serde_json::to_string(&edu).unwrap(),
             r#"{"edu_type":"m.device_list_update","content":{"user_id":"@john:example.com","device_id":"QBUAZIFURK","device_display_name":"Mobile","stream_id":6,"deleted":false,"keys":{"user_id":"@alice:example.com","device_id":"JLAFKJWSCS","algorithms":["m.olm.v1.curve25519-aes-sha2","m.megolm.v1.aes-sha2"],"keys":{"curve25519:JLAFKJWSCS":"3C5BFWi2Y8MaVvjM8M22DBmh24PmgR0nPvJOIArzgyI","ed25519:JLAFKJWSCS":"lEuiRJBit0IG6nUf5pUzWTUEsRVVe/HJkoKuEww9ULI"},"signatures":{"@alice:example.com":{"ed25519:JLAFKJWSCS":"dSO80A01XiigH3uBiDVx/EjzaoycHcjq9lfQX0uWsqxl2giMIiSPR8a4d291W1ihKJL/a+myXS367WT6NAIcBA"}}}}}"#
+        );
+    }
+
+    #[test]
+    fn minimal_device_list_update_edu() {
+        let json = json!({
+                "content": {
+                    "device_id": "QBUAZIFURK",
+                    "stream_id": 6,
+                    "user_id": "@john:example.com"
+                },
+            "edu_type": "m.device_list_update"
+        });
+
+        let edu = serde_json::from_value::<Edu>(json).unwrap();
+        assert_matches!(
+            &edu,
+            Edu::DeviceListUpdate(DeviceListUpdateContent {
+                user_id,
+                device_id,
+                device_display_name,
+                stream_id,
+                prev_id,
+                deleted,
+                keys
+            }) if user_id == &user_id!("@john:example.com")
+                && device_id == &device_id!("QBUAZIFURK")
+                && device_display_name.is_none()
+                && stream_id == &uint!(6)
+                && prev_id.is_empty()
+                && deleted.is_none()
+                && keys.is_none()
+        );
+
+        assert_eq!(
+            serde_json::to_string(&edu).unwrap(),
+            r#"{"edu_type":"m.device_list_update","content":{"user_id":"@john:example.com","device_id":"QBUAZIFURK","stream_id":6}}"#
         );
     }