+++ crates/ruma-federation-api/src/lib.rs
@@ -17,7 +17,6 @@ pub mod directory;
 pub mod discovery;
 pub mod event;
 pub mod keys;
-#[cfg(feature = "unstable-pre-spec")]
 pub mod knock;
 pub mod membership;
 pub mod openid;